Refactor container restart policy tests to e2e/common/node
- Added validation for lastTerminationStatus
